correction de bug l102 bordereau.php

+++ Pages/Bordereau/bordereau2.php
@@ -94,7 +94,9 @@
                     <th class="thBordereau">Total</th>
                     <?php 
                     if($_SESSION['role'] == "tresorier"){
-                        echo '<th class="thBordereauIcon">Action</th>';
+                        if ($bordereau->get_validiteTresorier() == 0){
+                            echo '<th class="thBordereauIcon">Action</th>';
+                        }
                     }
                     ?>
                 </tr>
@@ -119,8 +121,10 @@
                         echo "<td>".$ligneDeFrais->get_coutHebergement()."</td>";
                         echo "<td>".$ligneDeFrais->get_coutTotal()."</td>";
                         if($_SESSION['role'] == "tresorier"){
-                            echo '<td><a href="..\Action\edit.php?id=' . $ligneDeFrais->get_idFrais() .'&amp;email='.$email.'"><img id="edit" src="../../ico/edit.png"/></a> '
-                                . '<a href="..\Action\delete.php?id=' . $ligneDeFrais->get_idFrais() .'&amp;email='.$email.'"><img id="delete" src="../../ico/del.png"/></a></td>';
+                            if ($bordereau->get_validiteTresorier() == 0){
+                                echo '<td><a href="..\Action\edit.php?id=' . $ligneDeFrais->get_idFrais() .'&amp;email='.$email.'"><img id="edit" src="../../ico/edit.png"/></a> '
+                                    . '<a href="..\Action\delete.php?id=' . $ligneDeFrais->get_idFrais() .'&amp;email='.$email.'"><img id="delete" src="../../ico/del.png"/></a></td>';
+                            }
                         }
                         echo '</tr>';
                     }
